Ignore errors when a tramp method lacks of param as static value
`tramp-get-method-parameter` first tries to get the parameter from the
connection properties, afterwards tries from the static value defined in
`tramp-methods'. When a tramp method lacks of the param appropriately returns
nil; but pass to the branch used for compatibility old TRAMP versions, in this
case is incorrectly called with the method name instead of a tramp
vector. e.g. "sshx" method.

Closes #12